Use int32 instead of int

+++ exported.go
@@ -20,27 +20,27 @@ import ()
 // The type representation of a Monotonic Gray Code number
 type Mgc int
 
-var convert_to_int []int
+var convert_to_int []int32
 var convert_to_mgc []Mgc
 
 // Initialize a sequence of numbers, of length 'count'
 // The tables are pre computed to speed up conversion
 func Init(count int) {
-	convert_to_int = make([]int, count)
+	convert_to_int = make([]int32, count)
 	convert_to_mgc = make([]Mgc, count)
 	m := monotonic(count)
 	for i, mgc := range m {
-		convert_to_int[getValue(mgc)] = i
+		convert_to_int[getValue(mgc)] = int32(i)
 		convert_to_mgc[i] = Mgc(getValue(mgc))
 	}
 }
 
 // Convert a Monotonic Gray Code to the corresponding integer. Init() has to called first.
-func GetInt(n Mgc) int {
+func GetInt(n Mgc) int32 {
 	return convert_to_int[n]
 }
 
 // Get the n:th Monotonic Gray Code. Init() has to called first.
-func GetMgc(n int) Mgc {
+func GetMgc(n int32) Mgc {
 	return convert_to_mgc[n]
 }
